// src/entities/calculate/model/types.ts
export type LightingStandard = "SNIP" | "EU";

export type RoomType =
  | "work_office"
  | "office"
  | "computer_room"
  | "classroom"
  | "bank_operating_room"
  | "reading_room"
  | "design_office"
  | "conference_room"
  | "sports_hall"
  | "exhibition_hall"
  | "retail_sales_area"
  | "dining_room"
  | "hair_salon"
  | "doctor_office"
  | "garage"
  | "warehouse_receiving"
  | "warehouse_storage"
  | "lobby"
  | "corridor"
  | "stairs"
  | "attic";

export type FixtureType =
  | "recessed"
  | "track"
  | "modular"
  | "wall_ceiling"
  | "linear"
  | "suspended"
  | "linear_led"
  | "information_display";

export type CeilingType = "white" | "light" | "dark";
export type WallType = "light" | "dark";
export type FloorType = "dark" | "black";